ContractCoverageWriter
